removed helper pages
