+++ components/layout/AdminTools/index.js
@@ -1,0 +1,2 @@
+export * from './AdminTools';
+export { default } from './AdminTools';
